requirements.txt
pandas
ipykernel
openpyxl
python-dotenv